Bugfix: Null values crashing Orm database accesses.
When poking a null value into a PreparedStatement, the Orm class was not using
setNull(). This resulted in NPEs when using auto-unboxing (for Integers and
setInt(), for example).

There is a similar problem with named queries, but that will require a
configuration change to make it work (need to know whether the param is
nullable, and what its SQL type is for setNull()).

It would be nice to check whether non-nullable values are being set to null,
and to throw a descriptive NPE in that case, but we cannot test null against
primitive types.