Bugfix: SQLite does not support multiple-insert.
The example Person class used a 'friendmap' table of pairs of person IDs, with
each pair of friends appearing in the table twice, in both orders. But since it
is not possible to insert both rows in the same statement with SQLite, it needs
to do things differently.

For SQLite only, we insert just one row, and use a Union select to grab the
friend's ID regardless of the ordering. As a bonus, this shows another use
of dialects.