Bugfix: Calling SormContext.getSession() should never close a session.
It was closing an existing session if the cache mode was incorrect, which
did not make sense, as not providing a cache mode means you do not really
care what the cache mode is, and therefore should not change it.

Detected during testing when the populate() function called getSession(),
causing the existing one (referred to by the caller) to close. Then, when
the caller use its session, it failed because it was closed.