Add inheritance.
This is fairly simple and straightforward -- The configuration may include
a reference to a "super" class, which the generated class extends.
Fields may be configured as 'from-super' if they do not actually exist on
the subclass, allowing them to be referenced by SQL statements. Getters and
setters may be marked as 'override' to add @Override annotations.