Allow for grouped fields.
A grouped field is a single field on a class that is represented by
multiple columns in the database, on the same table as the rest of the
object. For example, the (x, y, z) coordinates of a point in 3-space could
be represented by a 3-dimensional vector object in Java, but in a
database you would want to just have the x, y, and z coordinates directly
written to the same row as the rest of the object, as the only other choice
would be to link it somewhere, and links must be nullable.

Group parents have group=true and no parent. They have a translation
snippet to go from SQL to Java (by building the conglomerate object out
of the database fields), but no translation snippet for going from Java to
SQL, as it would need to provide multiple columns worth of data.

Group children (subfields) have group=false, but they provide the name of
the parent field. These fields do not have a translation snippet from SQL
to Java, as they never exist on their own. Similarly, they do not get
accessors on the generated class. They do have Java to SQL translation
snippets, so that each one may be pulled out of the conglomerate object
individually when writing to the database.

Still need unit tests for this, especially to make sure that subfield
translation snippets are used when writing to the database.