Parser-side: Add options for generating Sorm object identifiers.
Current options are Post (the default), which calls the primary key getter
after inserting the object into the database, and Pre, which calls the
primary key getter before inserting the object.

The goal with Pre is to separate object insertion from primary key
generation, even if the primary key is generated by a method other than
simply inserting the object.

Still need to implement the feature in the code generator.