Support for pre- and post-insert ID generators and multi-statement queries.
The pre-insert ID generator allows for inserting into a master table of
generic objects, for example, and using that ID when inserting into the
specific object's table. It requires being able to execute multiple SQL
statements as part of a single query.

Multiple statements in a single query will split on semicolon and run all
but the last statement as updates, with the last coming back as a
PreparedStatement that can be used for a query (i.e. for primary key
getters).

Unfortunately, you cannot currently use any properties of the object in the
update statements, only in the PreparedStatement at the end. Need to
heavily reorganize the code generation to make complete PreparedStatement
generation and execution a single module.